Move types to separate file
